+++ src/web/api_handlers_recordings_list.c
@@ -263,14 +263,14 @@ void handle_get_recordings(const http_request_t *req, http_response_t *res) {
         if (recordings[i].size_bytes < 1024) {
             snprintf(size_str, sizeof(size_str), "%ld B", recordings[i].size_bytes);
         } else if (recordings[i].size_bytes < (uint64_t)1024 * 1024) {
-            snprintf(size_str, sizeof(size_str), "%.1f KB", recordings[i].size_bytes / 1024.0);
+            snprintf(size_str, sizeof(size_str), "%.1f KB", (double)recordings[i].size_bytes / 1024.0);
         } else if (recordings[i].size_bytes < (uint64_t)1024 * 1024 * 1024) {
-            snprintf(size_str, sizeof(size_str), "%.1f MB", recordings[i].size_bytes / (1024.0 * 1024.0));
+            snprintf(size_str, sizeof(size_str), "%.1f MB", (double)recordings[i].size_bytes / (1024.0 * 1024.0));
         } else {
-            snprintf(size_str, sizeof(size_str), "%.1f GB", recordings[i].size_bytes / (1024.0 * 1024.0 * 1024.0));
+            snprintf(size_str, sizeof(size_str), "%.1f GB", (double)recordings[i].size_bytes / (1024.0 * 1024.0 * 1024.0));
         }
 
-        cJSON_AddNumberToObject(recording, "id", recordings[i].id);
+        cJSON_AddNumberToObject(recording, "id", (double)recordings[i].id);
         cJSON_AddStringToObject(recording, "stream", recordings[i].stream_name);
         cJSON_AddStringToObject(recording, "file_path", recordings[i].file_path);
         cJSON_AddStringToObject(recording, "start_time", start_time_formatted);
